+++ packages/clients/diplan/src/config.js
@@ -12,6 +12,11 @@ export default {
     248651.73157077, 5227198.20287631, 928366.12236557, 6118661.62507136,
   ],
   diplan: {
+    link: {
+      href: 'https://example.com',
+      icon: '$vuetify.icons.fullscreen-exit',
+      label: 'diplan.linkButton.label',
+    },
     mergeToMultiGeometries: true,
     validateGeoJson: true,
     metaServices: [
@@ -152,7 +157,6 @@ export default {
     ],
   },
   draw: {
-    enableOptions: true,
     lassos: [
       {
         id: flurstuecke,
@@ -161,19 +165,7 @@ export default {
         id: xplanwfs,
       },
     ],
-    measureOptions: {
-      metres: true,
-      kilometres: true,
-      hectares: true,
-    },
-    selectableDrawModes: ['Point', 'LineString', 'Circle', 'Text', 'Polygon'],
     snapTo: [xplanwfs, flurstuecke],
-    textStyle: {
-      font: {
-        size: [10, 20, 30],
-        family: 'Arial',
-      },
-    },
     style: {
       fill: { color: 'rgb(51 117 212 / 50%)' },
       stroke: {
@@ -185,9 +177,6 @@ export default {
         fillColor: 'rgb(51 117 212 / 50%)',
       },
     },
-  },
-  export: {
-    displayComponent: false,
   },
   gfi: {
     mode: 'bboxDot',
@@ -207,13 +196,5 @@ export default {
       'plugin/pins/transformedCoordinate',
       'plugin/pins/coordinatesAfterDrag',
     ],
-  },
-  pins: {
-    toZoomLevel: 9,
-    movable: 'drag',
-    appearOnClick: {
-      show: true,
-      atZoomLevel: 0,
-    },
   },
 }