+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
 import * as mpapi from '@masterportal/masterportalapi'
-import addPlugins from './utils/addPlugins'
+import addPlugins, { resetPlugins } from './utils/addPlugins'
 import createMap from './utils/createMap'
 
 // NOTE: This is needed to be able to properly use the export
@@ -16,12 +16,14 @@ export type { MapInstance } from './types'
 export type PolarCore = typeof mpapi & {
   // TODO add more
   addPlugins: typeof addPlugins
+  resetPlugins: typeof resetPlugins
   createMap: typeof createMap
 }
 
 const core: PolarCore = {
   ...mpapi,
   addPlugins,
+  resetPlugins,
   createMap,
 }