+++ packages/core/src/utils/addPlugins.ts
@@ -1,5 +1,10 @@
+let initialCreateMap = null
+
 export default function addPlugins(this, plugins) {
   const originalCreateMap = this.createMap
+  if (!initialCreateMap) {
+    initialCreateMap = originalCreateMap
+  }
   this.createMap = async (params) => {
     try {
       const instance = await originalCreateMap(params)
@@ -8,5 +13,12 @@ export default function addPlugins(this, plugins) {
     } catch (error) {
       console.error('@polar/core: Map creation failed.', error)
     }
+  }
+}
+
+export function resetPlugins(this) {
+  if (initialCreateMap) {
+    this.createMap = initialCreateMap
+    initialCreateMap = null
   }
 }